Move addition of table representations to its own file
