+++ stellarphot/settings/models.py
@@ -6,7 +6,6 @@ from pathlib import Path
 from typing import Annotated, Any, Literal, Optional, TypeVar
 
 from astropy.coordinates import EarthLocation, Latitude, Longitude, SkyCoord
-from astropy.io.misc.yaml import AstropyDumper, AstropyLoader
 from astropy.time import Time
 from astropy.units import Quantity, Unit, UnitConversionError
 from astropy.utils import lazyproperty
@@ -126,36 +125,6 @@ class BaseModelWithTableRep(BaseModel):
 
     def __init__(self, *arg, **kwargs):
         super().__init__(*arg, **kwargs)
-        self.generate_table_representers()
-
-    def generate_table_representers(self):
-        """
-        Call this method during initialization of a class to add the YAML
-        Table representation for the class.
-        """
-        class_string = f"!{self.__class__.__name__}"
-
-        # Add YAML round-tripping for the model
-        def _representer(dumper, model):
-            # THIS SHOULD TAP INTO ASTROPY'S YAML DUMPER SOMEHOW
-            # This is a little hacky at the moment. It seems like YAML
-            # has trouble reading in a dictionary, so though model.model_dump()
-            # works fine for writing, we can't construct from the dump.
-            #
-            # Instead of figuring out the right way to do that, this just dumps
-            # the json representation as a string.
-            return dumper.represent_mapping(
-                class_string, {"model_json_string": model.model_dump_json()}
-            )
-
-        def _constructor(loader, node):
-            # This loads the simple dictionary we dumped in _representer,
-            # then initializes the model with the json string.
-            mapping = loader.construct_mapping(node)
-            return self.__class__.model_validate_json(mapping["model_json_string"])
-
-        AstropyDumper.add_representer(self.__class__, _representer)
-        AstropyLoader.add_constructor(class_string, _constructor)
 
 
 class Camera(BaseModelWithTableRep):
